requirements.txt
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
maxminddb==2.6.0
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.224.1
typing_extensions==4.10.0
uvicorn==0.29.0
